client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.11.0",
    "docker>=7.1.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "tomli>=2.3.0",
]
